customate/requirements.txt
aiohttp==3.5.4
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
boto3==1.9.122
botocore==1.12.122
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
colorclass==2.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
decorator==4.4.0
defusedxml==0.5.0
Django==2.1.7
django-appconf==1.0.3
django-cors-headers==2.5.2
django-cryptography==0.3
django-enumfields==1.0.0
django-filter==2.1.0
django-guardian==1.5.0
django-phonenumber-field==2.3.1
django-polymorphic==2.0.3
django-storages==1.7.1
djangorestframework==3.9.2
djangorestframework-jsonapi==2.7.0
docopt==0.6.2
docutils==0.14
drf-nested-routers==0.91
ecdsa==0.13
envs==1.3
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
ipython==7.4.0
ipython-genutils==0.2.0
isodate==0.6.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
jsonapi-client==0.9.7
jsonschema==3.0.1
jwt==0.6.1
lxml==4.3.3
Markdown==3.1
MarkupSafe==1.1.1
multidict==4.5.2
packaging==19.0
parso==0.3.4
pexpect==4.6.0
phonenumbers==8.10.8
pickleshare==0.7.5
pip-upgrader==1.4.6
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.8.0
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-jose-cryptodome==1.3.2
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-toolbelt==0.9.1
s3transfer==0.2.0
six==1.12.0
terminaltables==3.1.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
warrant==0.6.1
wcwidth==0.1.7
yarl==1.3.0
zeep==3.3.1
